requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-ckeditor==6.7.2
django-constance==4.3.2
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.15.2
drf-yasg==1.21.9
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Markdown==3.7
packaging==24.2
pillow==11.1.0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
